Remove old view code
